requirements.txt
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
Pillow==11.1.0
numpy==2.2.3
